refactor: change table plugin directory and file name
