+++ apps/editor/src/wysiwyg/plugins/selection/tableSelection.ts
