+++ apps/editor/src/wysiwyg/plugins/selection/tableSelectionView.ts
